Graffiti now has a game-wide limit per PC
